Run fmt, clippy, and test on all non-runner crates (#143)

+++ crates/prelude/src/crypto/ccm.rs
@@ -30,7 +30,7 @@ pub fn is_supported() -> bool {
 /// Returns the ciphertext given a cleartext, a key, and a nonce.
 pub fn encrypt(key: &[u8; 16], nonce: &[u8; 8], clear: &[u8]) -> Result<Vec<u8>, Error> {
     let len = clear.len();
-    let mut cipher = vec![0; len as usize + 4];
+    let mut cipher = vec![0; len + 4];
     let params = api::encrypt::Params {
         key: key.as_ptr(),
         iv: nonce.as_ptr(),
@@ -46,7 +46,7 @@ pub fn encrypt(key: &[u8; 16], nonce: &[u8; 8], clear: &[u8]) -> Result<Vec<u8>,
 /// Returns the cleartext given a ciphertext, a key, and a nonce.
 pub fn decrypt(key: &[u8; 16], nonce: &[u8; 8], cipher: &[u8]) -> Result<Vec<u8>, Error> {
     let len = cipher.len() - 4;
-    let mut clear = vec![0; len as usize];
+    let mut clear = vec![0; len];
     let params = api::decrypt::Params {
         key: key.as_ptr(),
         iv: nonce.as_ptr(),
